ajust database schema for groups

+++ Modules/Groups/Database/Migrations/2018_12_14_132158_create_groups_table.php
@@ -26,10 +26,12 @@ class CreateGroupsTable extends Migration
             $table->timestamps();
         });
 
-        Schema::table('wishes', function (Blueprint $table) {
-            $table->integer('group_id')->after('created_by')->nullable()->unsigned()->index();
-            $table->foreign('group_id')->references('id')->on('groups')->onDelete('cascade');
-        });
+        if (!Schema::hasColumn('wishes', 'group_id')) {
+            Schema::table('wishes', function (Blueprint $table) {
+                $table->integer('group_id')->after('created_by')->nullable()->unsigned()->index();
+                $table->foreign('group_id')->references('id')->on('groups')->onDelete('cascade');
+            });
+        }
 
         Schema::table('groups', function (Blueprint $table) {
             $table->foreign('created_by')->references('id')->on('users')->onDelete('cascade');
